refactor: 音声合成 APIs をモジュール化
